rename / move blocks; add .ts to .vue
