+++ src/components/blocks/OneWireTempSensor/OneWireTempSensor.vue
