+++ src/components/blocks/SetPointSimple/CreateSetPointSimple.vue
