+++ src/components/blocks/SetPointSimple/SetPointSimple.vue
