+++ src/components/grid/GridContainer.vue
@@ -1,7 +1,8 @@
+<script lang="ts">
 import Vue from 'vue';
 import Component from 'vue-class-component';
 
-import GridItem from './grid-item.vue';
+import GridItem from './GridItem.vue';
 
 /* eslint-disable indent */
 @Component({
@@ -120,3 +121,43 @@ export default class GridContainer extends Vue {
     );
   }
 }
+</script>
+
+<style scoped>
+.grid-container {
+  background-color: transparent;
+}
+
+.grid-main-container, .grid-container-overlay {
+  position: relative;
+  display: grid;
+  grid-gap: 20px;
+  grid-template-columns: repeat(auto-fill, 100px);
+  grid-auto-columns: 100px;
+  grid-auto-rows: 100px;
+  justify-content: center;
+}
+
+.grid-container-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  grid-template-rows: auto;
+}
+
+.grid-container-overlay-grid {
+  background-size: 120px 120px;
+  background-image:
+    linear-gradient(#121a1f 20px, transparent 0px),
+    linear-gradient(90deg, #121a1f 20px, transparent 0px),
+    linear-gradient(#fff, #fff);
+  background-position: 0 -20px, -20px, 0 0;
+  grid-column-start: 1;
+  grid-column-end: -1;
+  grid-row-start: 1;
+  grid-row-end: 1 span;
+  opacity: 0.03;
+}
+</style>
