+++ src/components/settings/Controllers.vue
@@ -1,3 +1,49 @@
+<script lang="ts">
+import Vue from 'vue';
+import Component from 'vue-class-component';
+
+import { controllers, isFetching } from '@/store/settings/getters';
+import { addController, removeController } from '@/store/settings/actions';
+
+@Component
+class Controllers extends Vue {
+  controllerInput: string = '';
+  $q: any;
+
+  get isFetching() {
+    return isFetching(this.$store);
+  }
+
+  get controllers() {
+    return controllers(this.$store);
+  }
+
+  addController() {
+    addController(this.$store, this.controllerInput);
+
+    // reset input
+    this.controllerInput = '';
+  }
+
+  removeController(controller: string) {
+    removeController(this.$store, controller);
+  }
+
+  remove(controller: string) {
+    this.$q.dialog({
+      title: 'Remove',
+      message: `Do you want to remove controller '${controller}'?`,
+      ok: 'Yes',
+      cancel: 'Cancel',
+    })
+      .then(() => this.removeController(controller))
+      .catch(() => {});
+  }
+}
+
+export default Controllers;
+</script>
+
 <template>
   <q-card>
     <q-card-title>Connected controllers</q-card-title>
@@ -39,8 +85,6 @@
     </q-card-main>
   </q-card>
 </template>
-
-<script lang="ts" src="./controllers.ts" />
 
 <style scoped>
 .q-item-main form {
