+++ src/pages/settings.vue
@@ -1,0 +1,19 @@
+<script>
+import Controllers from '@/components/settings/Controllers.vue';
+
+export default {
+  name: 'Settings',
+  components: {
+    Controllers,
+  },
+};
+</script>
+
+<template>
+  <q-page padding>
+    <controllers />
+  </q-page>
+</template>
+
+<style scoped>
+</style>
