--- src/pages/settings/settings.vue
@@ -1,20 +1,0 @@
-<template>
-  <q-page padding>
-    <controllers />
-  </q-page>
-</template>
-
-<script>
-import Controllers from './controllers';
-
-export default {
-  name: 'Settings',
-  components: {
-    Controllers,
-  },
-};
-</script>
-
-<style scoped>
-
-</style>
